+++ R/transactions.R
@@ -62,9 +62,9 @@ add_transactions <- function(.data, trx_data,
   # # column renames
   trx_data <- trx_data |>
     rename(pol_num = {{col_pol_num}},
-                  trx_date = {{col_trx_date}},
-                  trx_type = {{col_trx_type}},
-                  trx_amt = {{col_trx_amt}})
+           trx_date = {{col_trx_date}},
+           trx_type = {{col_trx_type}},
+           trx_amt = {{col_trx_amt}})
 
 
   # check for conflicting transaction types
@@ -80,7 +80,7 @@ add_transactions <- function(.data, trx_data,
   trx_data <- trx_data |>
     # between-join by transaction date falling within exposures windows
     inner_join(
-      date_lookup, multiple = "error",
+      date_lookup, relationship = "many-to-one",
       dplyr::join_by(pol_num,
                      between(trx_date, !!date_cols[[1]], !!date_cols[[2]])))
 
@@ -100,6 +100,6 @@ add_transactions <- function(.data, trx_data,
   .data |>
     left_join(trx_data, dplyr::join_by(pol_num, !!date_cols[[1]])) |>
     mutate(dplyr::across(dplyr::starts_with("trx_"), \(x)
-                                dplyr::coalesce(x, 0)))
+                         dplyr::coalesce(x, 0)))
 
 }